Move max sizes to powers of two
Magic numbers like `233` with only the nanopb being constrainted is strange.  This imposes a theoredical limit to what should be a `bytes` in the PB definition.  We should not use magic numbers, but instead use powers of two where artifical limits are imposed, and should target to have these limits be large as feesible for the firmware since its chopping messages.